Move plan data for code_8008C1D0 into asm files
Had to create a new file in order to get stuff matching right,
 even though making a separate file for just 3 functions is
 incredibly silly.
