+++ mk64.ld
@@ -102,6 +102,7 @@ SECTIONS
       BUILD_DIR/src/code_80071F00.o(.text);
       BUILD_DIR/src/code_80086E70.o(.text);
       BUILD_DIR/src/code_8008C1D0.o(.text);
+      BUILD_DIR/src/code_80091440.o(.text);
       BUILD_DIR/src/code_80091750.o(.text);
       BUILD_DIR/src/code_800AF9B0.o(.text);
       BUILD_DIR/src/menus.o(.text);
@@ -249,6 +250,9 @@ SECTIONS
       BUILD_DIR/src/code_80027D00.o(.data);
       BUILD_DIR/src/spawn_players.o(.data);
       BUILD_DIR/data/data_0DD0A0_2_0.o(.data);
+      BUILD_DIR/src/code_8008C1D0.o(.data);
+      BUILD_DIR/src/code_80091440.o(.data);
+      BUILD_DIR/data/data_0DD0A0_2_1.o(.data);
       BUILD_DIR/src/code_800AF9B0.o(.data);
       BUILD_DIR/src/menus.o(.data);
       BUILD_DIR/src/code_800B45E0.o(.data);
