+++ src/code_8008C1D0.c
@@ -338,10 +338,6 @@ void func_8008C8C4(Player* player, s8 arg1) {
     }
 }
 
-#ifdef NEEDS_RODATA
-// data_0DD0A0_2_0.s
-extern s16 D_800E6F60[];// = { 1092, 1092, 2184, 1638, 1820 }
-
 void func_8008C9EC(Player *player, s8 arg1) {
     s16 stackPadding1;
     s16 stackPadding2;
@@ -395,9 +391,6 @@ void func_8008C9EC(Player *player, s8 arg1) {
         }
     }
 }
-#else
-GLOBAL_ASM("asm/non_matchings/code_8008C1D0/func_8008C9EC.s")
-#endif
 
 void func_8008CDC0(Player* player, s8 arg1) {
     func_8008C354(player, arg1);
@@ -1734,27 +1727,6 @@ void func_8008FF08(Player *player, s8 playerId) {
     }
 }
 
-#ifdef NEEDS_RODATA
-// data_0DD0A0_2_0.s
-extern f32 D_800E6F6C[4];// = {   0.0f,    0.0f, -700.0f, 700.0f };
-extern f32 D_800E6F7C[4];// = { 700.0f, -700.0f,    0.0f,   0.0f };
-extern f32 D_800E6F8C[4];// = {   0.0f,    0.0f, -650.0f, 650.0f };
-extern f32 D_800E6F9C[4];// = { 650.0f, -650.0f,    0.0f,   0.0f };
-extern f32 D_800E6FAC[4];// = {   0.0f,    0.0f, -400.0f, 400.0f };
-extern f32 D_800E6FBC[4];// = { 400.0f, -400.0f,    0.0f,   0.0f };
-extern f32 D_800E6FCC[4];// = {   0.0f,    0.0f, -350.0f, 350.0f };
-extern f32 D_800E6FDC[4];// = { 350.0f, -350.0f,    0.0f,   0.0f };
-extern f32 D_800E6FEC[4];// = {   0.0f,    0.0f, -675.0f, 675.0f };
-extern f32 D_800E6FFC[4];// = { 675.0f, -675.0f,    0.0f,   0.0f };
-extern f32 D_800E700C[4];// = {   0.0f,    0.0f, -550.0f, 550.0f };
-extern f32 D_800E701C[4];// = { 550.0f, -550.0f,    0.0f,   0.0f };
-extern f32 D_800E702C[4];// = {   0.0f,    0.0f, -575.0f, 575.0f };
-extern f32 D_800E703C[4];// = { 575.0f, -575.0f,    0.0f,   0.0f };
-extern f32 D_800E704C[4];// = {  10.0f,  -10.0f, -575.0f, 575.0f };
-extern f32 D_800E705C[4];// = { 575.0f, -575.0f,   10.0f, -10.0f };
-// data_0DD0A0_3_0.s
-// jpt_800EF5F8
-
 void func_80090178(Player *player, s8 playerId, Vec3f arg2, Vec3f arg3) {
     u16 test;
     struct TrackWayPoint *temp_v1;
@@ -1832,9 +1804,6 @@ void func_80090178(Player *player, s8 playerId, Vec3f arg2, Vec3f arg3) {
         break;
     }
 }
-#else
-GLOBAL_ASM("asm/non_matchings/code_8008C1D0/func_80090178.s")
-#endif
 
 void func_80090778(Player* player) {
     s32 playerIndex = get_player_index_for_player(player);
@@ -2146,15 +2115,10 @@ void func_800911B4(Player* player, s8 arg1) {
     } while (temp_v0 < 10);
 }
 
-#ifdef NEEDS_RODATA
-// data_0DD0A0_2_0.s
-// May actually be a Vec3f? Its hard to tell if the 0.0f is part of the array or not
-extern f32 D_800E706C[4];// = { 27.16666666666f, 25.16666666666f, 23.16666666666f, 0.0f };
-
 void func_80091298(Player *player, s8 arg1) {
     s16 var_v1;
     s32 stackPadding1;
-    Vec3f spC = { 27.16666666666f, 25.16666666666f, 23.16666666666f };
+    Vec3f spC = { 27.167f, 25.167f, 23.167f };
 
     player->unk_044 |= 0x2000;
     if (player->unk_0B2 == 0) {
@@ -2179,7 +2143,7 @@ void func_80091298(Player *player, s8 arg1) {
                 player->unk_DC0 = 3.0f;
                 player->unk_044 &= ~0x800;
                 player->kartGravity = gKartGravityTable[player->characterId];
-                player->unk_0D4 = 0;
+                player->unk_0D4[0] = 0;
                 player->unk_000 |= 0x2000;
                 player->unk_094 = 0.0f;
                 player->unk_08C = 0.0f;
@@ -2191,95 +2155,3 @@ void func_80091298(Player *player, s8 arg1) {
         }
     }
 }
-#else
-GLOBAL_ASM("asm/non_matchings/code_8008C1D0/func_80091298.s")
-#endif
-
-void func_80091440(s8 arg0) {
-    if ((gPlayers[arg0].unk_044 & 0x800) == 0) {
-        gPlayers[arg0].unk_044 |= 0x2400;
-        gPlayers[arg0].unk_000 &= ~0x2000;
-    }
-}
-
-void func_800914A0(void) {
-    D_80152308 = gPlayerOneCopy->unk_006 + 7;
-    if ((s32) D_8015F6F8 < D_80152308) {
-        D_80152308 -= D_8015F6F8;
-    }
-}
-
-#ifdef MIPS_TO_C
-//generated by m2c commit d9d3d6575355663122de59f6b2882d8f174e2355 on Dec-15-2022
-extern u32 D_8015F718;
-static ? D_800E7080;                                /* unable to generate initializer */
-static ? D_800E708C;                                /* unable to generate initializer */
-
-void func_800914E0(void) {
-    ? sp64;
-    ? sp58;
-    s32 sp48;
-    f32 sp38;
-    f32 sp34;
-    f32 sp30;
-    Player **var_t1;
-    f32 temp_f12;
-    f32 temp_f14;
-    f32 temp_f16;
-    f32 temp_f4;
-    f32 temp_f8;
-    s16 temp_a0;
-    s16 temp_a0_2;
-    s16 temp_v0;
-    u16 temp_t0;
-    void *temp_v0_2;
-
-    sp64.unk0 = (s32) D_800E7080.unk0;
-    sp64.unk4 = (s32) D_800E7080.unk4;
-    sp64.unk8 = (s32) D_800E7080.unk8;
-    sp58.unk0 = (s32) D_800E708C.unk0;
-    sp58.unk4 = (s32) D_800E708C.unk4;
-    sp58.unk8 = (s32) D_800E708C.unk8;
-    temp_a0 = D_8015F718 & 0xFFFFFF;
-    var_t1 = &gPlayerOneCopy;
-    sp48 = gSegmentTable[(u32) D_8015F718 >> 0x18] + temp_a0 + 0x80000000;
-    temp_t0 = D_80152308;
-    temp_v0 = (s16) gPlayerOneCopy->unk_006 - temp_t0;
-    cameras->up[0] = 0.0f;
-    cameras->up[2] = 0.0f;
-    cameras->up[1] = 1.0f;
-    if (temp_v0 == 7) {
-        func_800914A0(temp_a0, 7);
-        goto block_5;
-    }
-    if (temp_v0 < 0xE) {
-        temp_a0_2 = temp_v0 + (s16) D_8015F6F8;
-        if (temp_a0_2 == 7) {
-            func_800914A0(temp_a0_2, 7);
-block_5:
-            var_t1 = &gPlayerOneCopy;
-        }
-    }
-    temp_v0_2 = (temp_t0 * 8) + sp48;
-    temp_f8 = (f32) (temp_v0_2->unk0 + 0xA);
-    cameras->pos[0] = temp_f8;
-    temp_f16 = (f32) (temp_v0_2->unk2 + 7);
-    cameras->pos[1] = temp_f16;
-    temp_f4 = (f32) (temp_v0_2->unk4 - 0x14);
-    cameras->pos[2] = temp_f4;
-    cameras->lookAt[0] = (*var_t1)->pos[0];
-    cameras->lookAt[1] = (*var_t1)->pos[1];
-    cameras->lookAt[2] = (*var_t1)->pos[2];
-    func_802ADDC8(&cameras->unk_54, 20.0f, temp_f8, temp_f16, temp_f4);
-    temp_f12 = cameras->lookAt[0] - cameras->pos[0];
-    sp38 = temp_f12;
-    temp_f14 = cameras->lookAt[2] - cameras->pos[2];
-    sp34 = cameras->lookAt[1] - cameras->pos[1];
-    sp30 = temp_f14;
-    cameras->rot[1] = atan2s(temp_f12, temp_f14);
-    cameras->rot[0] = atan2s(sqrtf((sp38 * sp38) + (temp_f14 * temp_f14)), sp34);
-    cameras->rot[2] = 0;
-}
-#else
-GLOBAL_ASM("asm/non_matchings/code_8008C1D0/func_800914E0.s")
-#endif
