+++ src/code_80091440.h
@@ -1,0 +1,8 @@
+#include "ultra64.h"
+
+void func_80091440(s8);
+void func_800914A0();
+void func_800914E0();
+
+extern u16 D_80152308;
+extern u16 D_8015F6F8;
